replace remote with ipc
