+++ src/main/utils/ipcComms.ts
@@ -1,6 +1,7 @@
 import { BrowserWindow, dialog, ipcMain, Tray, Notification } from 'electron';
 import { createWindow } from './createWindow';
 import { buildContextMenu } from './menuUtil';
+import { getFilesFromDir } from './upload';
 
 export default function setupIpcComs(
     tray: Tray,
@@ -42,5 +43,25 @@ export default function setupIpcComs(
         const secondWindow = createWindow();
         mainWindow.destroy();
         mainWindow = secondWindow;
+    });
+
+    ipcMain.on('show-upload-files-dialog', async (event) => {
+        const files = await dialog.showOpenDialog({
+            properties: ['openFile', 'multiSelections'],
+        });
+        event.returnValue = files.filePaths;
+    });
+
+    ipcMain.on('show-upload-dirs-dialog', async (event) => {
+        const dir = await dialog.showOpenDialog({
+            properties: ['openDirectory', 'multiSelections'],
+        });
+
+        let files: string[] = [];
+        for (const dirPath of dir.filePaths) {
+            files = files.concat(await getFilesFromDir(dirPath));
+        }
+
+        event.returnValue = files;
     });
 }
